progress toward neighbor list generation

+++ src/main/scala/com/noackexpected/algorithm5/cluster/package.scala
@@ -24,5 +24,5 @@ package object cluster {
   type NeighborList = List[ItemID]
   type NeighborLists = Map[ItemID, NeighborList]
 
-  type Distance = (ItemID, ItemID, Float)
+  type Distance = (ItemID, ItemID, Double)
 }
